+++ app/policies/federails/server/publishable_policy.rb
@@ -1,0 +1,11 @@
+module Federails
+  module Server
+    class PublishablePolicy < Federails::FederailsPolicy
+      class Scope < Scope
+        def resolve
+          raise NotImplementedError
+        end
+      end
+    end
+  end
+end