+++ CHANGELOG.md
@@ -35,8 +35,11 @@ Breaking changes should be prefixed by `[**BREAKING**]` (without the quotes), to
 - `Federails::Actor`: Add `.distant` scope to select distant actors
 - `Federails::Request`: Add `.dereference` method to... dereference an object
 - New feature: Federated entities. This allows model configuration to ease the process of creating Fediverse entities 
-  from local content, and database entries from Fediverse content
+  from local content, and database entries from Fediverse content. When configured:
+  - "Create" activities will be created on data creation 
 - Data transformer for Notes: `Federails::DataTransformer::Note`, to ease transforming local data to Fediverse Notes
+- Server: new "published" controller to render published `Federails::DataEntity` as federated object. This controller 
+  will answer to the `federated_url` generated for local content.
 
 ## [0.4.0] 2024-12-02